+++ src/components/scenes/Staking/EarnScene.tsx
@@ -456,8 +456,8 @@ export const EarnScene = (props: Props) => {
           isSearching={isSearching}
           searchText={searchText}
           sceneWrapperInfo={sceneWrapperInfo}
-          onStartSearching={handleStartSearching}
-          onDoneSearching={handleDoneSearching}
+          onFocus={handleStartSearching}
+          onCancel={handleDoneSearching}
           onChangeText={handleChangeText}
           onLayoutHeight={handleFooterLayoutHeight}
         />