+++ src/components/scenes/WalletListScene.tsx
@@ -227,8 +227,8 @@ export const WalletListScene: React.FC<Props> = props => {
           searchText={searchText}
           noBackground
           sceneWrapperInfo={sceneWrapperInfo}
-          onStartSearching={handleStartSearching}
-          onDoneSearching={handleDoneSearching}
+          onFocus={handleStartSearching}
+          onCancel={handleDoneSearching}
           onChangeText={handleChangeText}
           onLayoutHeight={handleFooterLayoutHeight}
         />