+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - changed: Optimize login performance.
 - changed: Update Monero LWS server name to "Edge LWS"
 - fixed: ramps: Various Infinite UI/UX issues
+- fixed: Search keyboard not dismissing when submitting search
+- fixed: Auto-correct not disabled for search input
 
 ## 4.41.1 (2025-12-29)